renamed files to be all lowercase, and using camelCase
forgot to rename a few files, and updated some files to be able to successfully build the project
