+++ src/openSidepanel.js
